feat: make most of clippy happy and add ci to GitHub Actions (#116)
* fix: clippy needless_lifetimes


* fix: unnecessary_cast for clippy


* fix: useless_conversion for clippy


* fix: into_iter_on_ref for clippy


* fix: allow some


* fix: clone_on_copy clippy


* fix: comparison_to_empty clippy


* fix: len_zero clippy


* fix: needless_borrow clippy


* fix: needless_borrows_for_generic_args clippy


* fix: needless_question_mark clippy


* fix: nonminimal_bool clippy


* fix: collapsible_if clippy


* fix: unwrap_or_default clippy


* fix: redundant_static_lifetimes clippy


* fix: match_result_ok clippy


* fix: question_mark clippy


* fix: ineffective_open_options clippy


* fix: new allow


* fix: redundant_closure clippy


* fix: collapsible_else_if clippy


* fix: let_unit_value clippy


* fix

* fix: manual_flatten clippy


* fix: explicit_auto_deref clippy


* ci: add clippy to ci


* fix: update stable


---------

+++ src/utils/token_util.rs
@@ -11,9 +11,9 @@ use crate::{
 };
 
 // 24h
-pub const DEFAULT_LEASE_TTL: Duration = Duration::from_secs(24 * 60 * 60 as u64);
+pub const DEFAULT_LEASE_TTL: Duration = Duration::from_secs(24 * 60 * 60_u64);
 // 30d
-pub const MAX_LEASE_TTL: Duration = Duration::from_secs(30 * 24 * 60 * 60 as u64);
+pub const MAX_LEASE_TTL: Duration = Duration::from_secs(30 * 24 * 60 * 60_u64);
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct TokenParams {
